# api/openapi.yaml
openapi: 3.0.0
info:
  version: '0.1'
  title: System Configuration Service
  description: >-
    Commercial off-the-shelf BMCs and HPE-built controllers will need to have various
    parameters configured on them such as NTP server, syslog server, SSH keys, access
    credentials. These parameters are automatically configured when the targets like
    BMC and controllers are discovered.

    The System Configuration Service makes it possible for administrators to configure these
    parameters at anytime on multiple targets in a single operation. The System
    Configuration Service runs on the non-compute worker node.


    The REST API provides the following functions:

    * Set or retrieve network protocol (NWP) parameters (NTP, syslog servers, SSH keys)
    for specified targets (xnames or groups). Only Redfish access credentials can be configured
    for Commercial off-the-shelf BMCs or BMC's in air cooled hardware.
    The other parameters are not applicable.

    * Set Redfish access credentials for specified targets

    * Check service health

    ## Resources

    ### /bmc/dumpcfg

    Retrieve network protocol parameters like NTP, syslog server, and SSH keys for a
    set of controllers.

    ### /bmc/loadcfg

    Configure NTP, syslog server, and SSH keys for a set of controllers.

    ### /bmc/cfg/{xname}

    Retrieve or set NTP, syslog server, and SSH keys for a single controller.

    ### /bmc/discreetcreds

...

    ### /bmc/creds/{xname}

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
  /bmc/dumpcfg:
    post:
      tags:
        - nwp
        - cli_from_file
      summary: Retrieve the Redfish network protocol data for a set of targets
      description: >-
        Get the Redfish Network Protocol data (NTP server, syslog server, SSH
        key) for a set of targets.  The POST payload contains the parameters to
        retrieve  along with a list of targets.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bmc_dumpcfg_request'
      responses:
        '200':
          description: OK.  The data was succesfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bmc_dumpcfg_response'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only POST is allowed'
  /bmc/loadcfg:
    post:
      tags:
        - nwp
        - cli_from_file
      summary: Set the Redfish network protocol data for a set of targets
      description: >-
        Set the Redfish network protocol data (NTP server, syslog server, SSH
        key) for a set of targets. The POST payload contains the parameters to
        set  along with a list of targets.


        The Force field is optional.
        If present, and set to 'true', the Redfish operations will be attempted without contacting HSM
        and without verifying if the targets are present or are in a good state.
        If the "Force" field is not present or is present but set to 'false', HSM will be used.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bmc_loadcfg_request'
      responses:
        '200':
          description: OK.  The data was succesfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/multi_post_response'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only POST is allowed'
  '/bmc/cfg/{xname}':
    parameters:
      - in: path
        name: xname
        required: true
        schema:
          $ref: '#/components/schemas/xname'
      - in: query
        name: param
        schema:
          type: string
          anyOf:
            - $ref: '#/components/schemas/ntp_server_info_kw'
            - $ref: '#/components/schemas/syslog_server_info_kw'
            - $ref: '#/components/schemas/sshkey_kw'
            - $ref: '#/components/schemas/sshconkey_kw'
          description: 'Specification of network protocol parameters.'
    get:
      tags:
        - nwp
      summary: Retrieve Redfish network protocol data for a single target
      description: >-
        Retrieve selected Redfish network protocol data for a single target. You
        can select NTP server, Syslog server, or SSH key. If nothing is
        specified, all Redfish network protocol parameters are returned.
      responses:
        '200':
          description: OK.  The data was succesfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cfg_get_single'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only GET,POST is allowed'
    post:
      tags:
        - nwp
      summary: Set Redfish network protocol data for a single target
      description: >-
        Set selected Redfish network protocol data for a single target.
        Payload body specifies NTP server, syslog server, or SSH key.


        The Force field is optional. If present, and set to 'true', the Redfish operations
        will be attempted without contacting HSM
        and without verifying if the targets are present or are in a good state.
        If the "Force" field is not present or is present but set to 'false', HSM will be used.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/cfg_post_single'
      responses:
        '200':
          description: OK.  The data was succesfully set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cfg_rsp_status'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only GET,POST is allowed'
  /bmc/discreetcreds:
    post:
      tags:
        - creds
        - cli_from_file
      summary: Set the controller login credentials for a set of targets
      description: >-
        Set discrete controller login credentials for a set of targets.  The POST
        payload contains the parameters to set along with a list of targets.


        The Force field is optional. If present, and set to 'true', the Redfish operations
        will be attempted without contacting HSM
        and without verifying if the targets are present or are in a good state.
        If the "Force" field is not present or is present but set to 'false', HSM will be used.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/creds_components'
      responses:
        '200':
          description: OK.  The data was succesfully set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/multi_post_response'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only POST is allowed'
  '/bmc/creds/{xname}':
    parameters:
      - in: path
        name: xname
        required: true
        schema:
          $ref: '#/components/schemas/xname'
    post:
      tags:
        - creds
      summary: Set controller login credentials for a single target
      description: >-
        Set controller login credentials for a single target.  The POST
        payload contains the parameters to set along with a list of targets.


        The Force field is optional. If present, and set to 'true', the Redfish operations
        will be attempted without contacting HSM
        and without verifying if the targets are present or are in a good state.
        If the "Force" field is not present or is present but set to 'false', HSM will be used.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/creds_single'
      responses:
        '200':
          description: OK.  The data was succesfully set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cfg_rsp_status'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only GET,POST is allowed'
  /bmc/creds:
    parameters:
      - in: query
        name: targets
        required: false
        description: Comma separated list of XNames.  No query string results in fetching creds for all known BMCs.
        schema:
          $ref: '#/components/schemas/xname_list'
      - in: query
        name: type
        required: false
        description: Target component type.  A maximum of one type is allowed.  If no type is specified, all known BMC types are returned.
        schema:
          type: string
          example: NodeBMC
    get:
      tags:
        - creds
      summary: Fetch controller login credentials for specified targets and types.
      description: >-
        Fetch controller login credentials for a specified targets.  Targets
        are specified as a comma-separated list of xnames.  A component type
        may also be specified.  The xname list is 'ANDed' with the component
        type; any xname that has a type other than the specified type will
        be discarded.   If no type is specified, all BMC types are used.
        If no query parameters are specified at all, all BMCs in the system
        are used.
      responses:
        '200':
          description: OK.  The data was succesfully set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/creds_fetch_rsp'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only GET, POST is allowed'
  /bmc/globalcreds:
    post:
      tags:
        - creds
        - cli_from_file
      summary: Set the the same controller login credentials for a set of targets
      description: >-
        Set controller login credentials for a set of targets.  The POST
        payload contains the parameters to set along with a list of targets.
        The same credentials are set on all targets.


        The Force field is optional. If present, and set to 'true', the Redfish operations
        will be attempted without contacting HSM
        and without verifying if the targets are present or are in a good state.
        If the "Force" field is not present or is present but set to 'false', HSM will be used.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/creds_global'
      responses:
        '200':
          description: OK.  The data was succesfully set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/multi_post_response'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only POST is allowed'
  /bmc/createcerts:
    post:
      tags:
        - certs
        - cli_from_file
      summary: Create TLS cert/key pairs for a set of targets
      description: >-
        Create TLS cert/key pairs for a set of BMC targets.  A TLS cert/key
        is created per BMC 'domain', the default being one cert per cabinet
        to be used by all BMCs in that cabinet.  TLS cert/key info is stored
        in secure storage for subsequent application or viewing.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bmc_managecerts_request'
      responses:
        '200':
          description: OK.  The data was succesfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bmc_managecerts_response'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only POST or DELETE is allowed'
  /bmc/deletecerts:
    post:
      tags:
        - certs
        - cli_from_file
      summary: Delete TLS cert/key pairs for a set of targets
      description: >-
        Delete TLS cert/key information for domain-level TLS certs based on
        the given targets.  There will be one TLS cert/key per BMC 'domain'
        which will be deleted from secure storage.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bmc_managecerts_request'
      responses:
        '200':
          description: OK.  The data was succesfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bmc_managecerts_response'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only POST or DELETE is allowed'
  /bmc/fetchcerts:
    post:
      tags:
        - certs
        - cli_from_file
      summary: Fetch previously created BMC TLS certs for viewing.
      description: >-
        Fetches BMC TLS certs previously created using the /bmc/createcerts
        endpoint and stored in secure storage.  This API does not interact
        with Redfish BMCs.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bmc_managecerts_request'
      responses:
        '200':
          description: OK.  The data was succesfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bmc_fetchcerts_response'
        '404':
          description: Endpoint not found
        '405':
          description: 'Invalid method, only POST is allowed'
...
components:
  schemas:
...
    bmc_dumpcfg_request:
      type: object
      required:
        - Targets
      properties:
        Targets:
          type: array
          items:
            $ref: '#/components/schemas/xname'
        Params:
          type: array
          items:
            $ref: '#/components/schemas/cfg_types'
...
    bmc_dumpcfg_response:
      type: object
      properties:
        Targets:
          type: array
          items:
            $ref: '#/components/schemas/target_cfg_item'
...